Initial commit
--------------------------------------------------
Version:   EventBridge-Analysis-V6
--------------------------------------------------

Change Summary
--------------
This project contains source code and supporting files for a serverless application that you can deploy with the SAM CLI.

IMPORTANT
---------
This is a modified version of the code described in the README.MD.

The code has be modified structurally and extended. A Document has been generated to guide a developer through the process of building and tearing-down this code.

For this code to function, and AWS account must be available.
Experience with AWS: Lambda, API Gateway, S3, CloudWatch, CloudFormation is helpful.

These products should be installed:

        Docker Desktop and VSCode must be installed.
        AWS Toolkit, for VSCode is required.
        AWS CLI Configure, for VSCode is required.
        YAML from Read Hat, for VSCode.